Added platform prefab, changes to jump script
Changed grounded to public on jump script so I could access with with platform script. used [hidden] command to ensure it wasn't on the inspector.

Made changes to EnemyHealth manager to change it from creating a new object to instead just changing the sprite and then destroying it self. I am not sure if this is the best option.

Added StayOnPlatforms script to main character.